IBUs now rely on the grain bill

+++ OriginalGravityCalculator.scala
@@ -1,5 +1,5 @@
 object OriginalGravityCalculator {
-  def originalGravity(recipe: Recipe): BigDecimal = {
+  def originalGravity(recipe: Recipe): Double = {
     val potentialPoints = recipe.grainAdditions.map(_.potentialPoints).reduce(_ + _)
     val convertedPoints = potentialPoints * recipe.brewhouseEfficiency
     1 + (convertedPoints / recipe.batchVolume.inGallons) / 1000
